Reducing No. of Bands (Python)
Created python versions of the greenwood equation code, as well as my own reduceFilterNum function, in order to reduce the number of spectrogram and aperiodicity bands via python. This means that the size of the stored data remains small from the beginning. Then, in matlab solely resampling needs to be done to complete the preprocessing stage.